gcompat: Teach qq to accept both str and unicode + emit printable UTF-8 as is
This patch made its first step as a way to teach qq to also work on
python3. However the differences in str / unicode and escapes in between
py2 / py3 quickly popped out and then it became easier to just handle
whole escaping logic myself.

The implementation is based on

	https://lab.nexedi.com/kirr/go123/commit/c0bbd06e#d04d91f6343eea840875452d848aa5f8e342c895_0_46

and byproduct of manual handling is that now we don't escape printable
UTF-8 characters.