context: Fix thinko/typo in PyContext.__dealloc__
Instead of `pyctx.ctx = nil` it was just `ctx = nil` - i.e. assign nil
to local variable instead of changing pyctx instance data. We were not
observing this bug because Cython, for C++ fields of cdef classes,
automatically emits in-place destructor calls in generated __dealloc__

    https://github.com/cython/cython/blob/0.29.14-11-g8c620c388/Cython/Compiler/ModuleNode.py#L1477-L1478

and so this way there was no leak. However we want to be explicit and the
code was not correct. Fix it.

The bug was there from 2a359791 (context: Move/Port context package to
C++/Pyx nogil).