libgolang: tests: Factor out common testing functionality into shared place
Currently libgolang_test.cpp contains tests for code in libgolang.cpp and
for code that lives in other libgolang packages - sync, fmt, etc. It is
becoming tight and we are going to split libgolang_test.cpp and move
package tests to their corresponing files - e.g. to sync_test.cpp and
the like.

Move common assertion utilities into shared header before that as a
preparatory step.