golang: tests: Don't hide output details in assert on retcode of spawned command
For example if test_defer_excchain_dump_ipython fails the output was:

    >       assert retcode == 0
    E       assert 1 == 0

and it was unclear what was going on. Now the output is e.g.

    >       assert retcode == 0, (stdout, stderr)
    E       AssertionError: ('', '==152924==ASan runtime does not come first in initial library list; you should either link runtime to your application or manually preload it with LD_PRELOAD.
    E         ')
    E       assert 1 == 0

Amends: 09629367 (golang: tests: Add tests for IPython and Pytest
integration patches) and bb9a94c3 (golang: Teach defer to chain
exceptions (PEP 3134) even on Python2).