gpython: Rework pymain interface to accept python exe as argv[0]
In other words pymain now requires to be passed in full argv when
called. We will need exe=argv[0] in a followup patch to be able to
reexecute underlying python interpreter when given options like -O, -S,
etc...

This is change in behaviour and in general must come with changing
pymain name - e.g. to pymain2 - so that old pymain continues to work as
before. However since there are not so many known pymain users, we can
probably do this change in place since we'll care ourselves about all
those users:

- [python-interpreter] in SlapOS:
  https://lab.nexedi.com/nexedi/slapos/blob/46ed2afe/component/pygolang/buildout.cfg#L30-53

- pyruntraced in Go123:
  https://lab.nexedi.com/kirr/go123/blob/96046edf/tracing/cmd/pyruntraced

- zobjtrace in ZODB:
  https://lab.nexedi.com/kirr/ZODB/blob/07c21671/zobjtrace

So I hope it is ok.